==> notebooks/02_scrape_ita_links.ipynb <==
import pandas as pd
from scraper.scrape import fetch_html_for_urls
import bs4 as bs

data = pd.read_csv('data/unctad_clean.csv')

data = fetch_html_for_urls(data, url_col='link_to_italaws_case_page', html_col='italaw_html')

data.to_csv('data/unctad_clean_with_html.csv', index=False)
