==> notebooks/03_pull_data_from_pages.ipynb <==
import pandas as pd
from scraper.scrape import extract_titles, extract_case_metadata, attach_documents_to_data_list
import json
from bs4 import BeautifulSoup

data = pd.read_csv('data/unctad_clean_with_html.csv')

data = extract_titles(data, html_col='italaw_html', title_col='italaw_title')

fields = [
    ('italaw_case_type', 'views-field-field-case-type'),
    ('italaw_arbitration_rules', 'views-field-field-arbitration-rules'),
    ('italaw_investment_treaty', 'views-field-field-case-treaties'),
    ('italaw_legal_instruments', 'views-field-field-case-law-text'),
    ('italaw_economic_sector', 'views-field-field-economic-sector')
]
data = extract_case_metadata(data, html_col='italaw_html', fields=fields)

data_list = data.to_dict(orient='records')

data_list = attach_documents_to_data_list(data_list)


with open('data/unctad_clean_with_metadata.json', 'w') as f:
    json.dump(data_list, f, indent=4, ensure_ascii=False)
