==> notebooks/04_make_long_data.ipynb <==
import json
import pandas as pd
from utility.cleaning import flatten_to_document_level, to_snake_case

with open('data/unctad_clean_with_metadata.json', 'r') as f:
    data_list = json.load(f)

df = flatten_to_document_level(data_list)

del df['italaw_html']

df['short_case_name_clean'] = df['short_case_name'].apply(
    lambda x: to_snake_case(x) if pd.notna(x) else None
)

df['arbitration_id'] = df.apply(
    lambda row: f"{int(row['year_of_initiation'])}_{row['short_case_name_clean']}"
    if pd.notna(row['year_of_initiation']) and pd.notna(row['short_case_name_clean'])
    else None,
    axis=1
)


df['doc_name_clean'] = df['doc_name'].apply(
    lambda x: to_snake_case(x) if pd.notna(x) else None
)

df['doc_id'] = df.apply(
    lambda row: f"{row['arbitration_id']}_{row['doc_name_clean']}"
    if pd.notna(row['arbitration_id']) and pd.notna(row['doc_name_clean'])
    else None,
    axis=1
)

df.to_csv('data/unctad_document_level_data.csv', index=False, encoding='utf-8-sig')
