==> notebooks/05_download_docs.ipynb <==
import pandas as pd
from doc_download.download_docs import parallel_download_pdfs
import os


data = pd.read_csv('data/unctad_document_level_data.csv')

os.makedirs("documents", exist_ok=True)

results = parallel_download_pdfs(data)


from collections import Counter
Counter([r["status"] for r in results])

import json
with open("data/download_results.json", "w") as f:
    json.dump(results, f, indent=2)
